1.6.1: Variables and more metacommands
New variable construct: allows simple data to be stored per-server, deleted, and changed via commands. Retrieval is done using '$' followed by the variable name, which can be done by itself or within expressions. Certain variables are readable but not writable by the user, such as $server. Variables are case-insensitive, however they are cast to lowercase when assigning. Variables cannot hold data types other than string, number, and boolean. If a variable is not defined, either it will not insert its value, or a proper error will be thrown.

Example:
!batch {
	!var x 0;
	!repeat 5 {
		!var.inc x;
	};

	// or simply
	!foreach 5 x {!var x};

	!var.delete x;
}

Changes:
* New metacommands: var, foreach, while, cancel, and try
* Metacommands handle errors properly now
* Commands resolves rather than rejects
* CommandParser tokenization improvements:
  * Arguments cast to Number when possible
  * Expression evaluation done during normalization instead
* Arguments no longer throw errors if their value is defined but equivalent to false